Replace ellipses by circles: Rename files
Separate commit for file renames to make sure the renames are recognized
by Git.
