+++ libs/librepcb/common/dialogs/circlepropertiesdialog.ui
