+++ libs/librepcb/common/geometry/circle.cpp
@@ -21,7 +21,7 @@
  *  Includes
  ****************************************************************************************/
 #include <QtCore>
-#include "ellipse.h"
+#include "circle.h"
 
 /*****************************************************************************************
  *  Namespace
