+++ libs/librepcb/common/graphics/holegraphicsitem.h
@@ -25,7 +25,7 @@
  ****************************************************************************************/
 #include <QtCore>
 #include <QtWidgets>
-#include "primitiveellipsegraphicsitem.h"
+#include "primitivecirclegraphicsitem.h"
 #include "../geometry/hole.h"
 
 /*****************************************************************************************
