+++ libs/librepcb/libraryeditor/pkg/fsm/packageeditorstate_select.cpp
@@ -27,12 +27,12 @@
 #include <librepcb/library/pkg/package.h>
 #include <librepcb/library/pkg/footprintgraphicsitem.h>
 #include <librepcb/library/pkg/footprintpadgraphicsitem.h>
-#include <librepcb/common/graphics/ellipsegraphicsitem.h>
+#include <librepcb/common/graphics/circlegraphicsitem.h>
 #include <librepcb/common/graphics/polygongraphicsitem.h>
 #include <librepcb/common/graphics/stroketextgraphicsitem.h>
 #include <librepcb/common/graphics/holegraphicsitem.h>
 #include <librepcb/common/dialogs/polygonpropertiesdialog.h>
-#include <librepcb/common/dialogs/ellipsepropertiesdialog.h>
+#include <librepcb/common/dialogs/circlepropertiesdialog.h>
 #include <librepcb/common/dialogs/stroketextpropertiesdialog.h>
 #include <librepcb/common/dialogs/holepropertiesdialog.h>
 #include "cmd/cmdmoveselectedfootprintitems.h"
